python/flask-sqlalchemy/requirements.txt
flask-sqlalchemy==2.1
marshmallow